Modularize usx_to_usj
